refactor: Move WeatherCondition to constants directory
