+++ lib/domain/weather/constants/weather_condition.dart
